Moved TreeGridAxis and TreeGridTick.
