+++ ts/masters/modules/treegrid.src.ts
@@ -11,7 +11,7 @@
  */
 'use strict';
 import Highcharts from '../../Core/Globals.js';
-import TreeGridAxis from '../../Core/Axis/TreeGridAxis.js';
+import TreeGridAxis from '../../Core/Axis/TreeGrid/TreeGridAxis.js';
 const G: AnyRecord = Highcharts;
 // Compositions
 TreeGridAxis.compose(G.Axis, G.Chart, G.Series, G.Tick);
